功能更新
完善了由二进制数据包解析协议的功能
已完成Ethernet, IPv4, IPv6, ARP协议的解析
增加了十六进制数据的显示

+++ WinSniffer/WinSniffer/ProtocolAnalyzer/IPv4Analyzer.cs
@@ -16,7 +16,7 @@ namespace WinSniffer
         public int identification;
         public int flags;
         public int fragmentOffset;
-        public byte TTL;
+        public int TTL;
         public byte protocol;
         public int headerChecksum;
         public IPAddress sourceIP;
